src/main/java/com/esgglobal/customersapp/customer/CustomerController.java
package com.esgglobal.customersapp.customer;

import com.esgglobal.customersapp.customer.dto.Customer;
import lombok.AllArgsConstructor;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.web.bind.annotation.*;

@RestController
@RequestMapping(value = "/customers",
        produces = MediaType.APPLICATION_JSON_VALUE)
@AllArgsConstructor
public class CustomerController {

    private CustomerService service;

    @PostMapping
    @ResponseStatus(HttpStatus.CREATED)
    public void createCustomer(@RequestBody Customer customer){
        service.saveCustomer(customer);
    }

    @GetMapping("/{id}")
    public Customer getCustomer(@PathVariable String id){
        return service.getCustomer(id);
    }

    @ExceptionHandler
    @ResponseStatus(HttpStatus.NOT_FOUND)
    public ErrorMessage handleCustomerNotFound(CustomerNotFoundException cnfe){
        return new ErrorMessage(cnfe.getMessage(), HttpStatus.NOT_FOUND.name());
    }

}
