src/test/java/com/esgglobal/customersapp/customer/CustomerControllerTest.java
package com.esgglobal.customersapp.customer;

import com.esgglobal.customersapp.customer.dto.Customer;
import com.fasterxml.jackson.databind.ObjectMapper;
import org.hamcrest.Matchers;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.junit.jupiter.MockitoExtension;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import org.springframework.boot.test.mock.mockito.MockBean;
import org.springframework.http.MediaType;
import org.springframework.test.web.servlet.MockMvc;

import java.util.UUID;

import static org.mockito.Mockito.when;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;

@WebMvcTest(CustomerController.class)
@ExtendWith(MockitoExtension.class)
public class CustomerControllerTest {

    @Autowired
    private MockMvc mockMvc;

    @MockBean
    private CustomerService customerService;

    @Test
    void shouldCreateCustomerSuccessfully() throws Exception {
        Customer customer = CustomerProducer.customerDto(UUID.randomUUID());
        this.mockMvc.perform(post("/customers")
                        .content(asJsonString(customer))
                        .contentType(MediaType.APPLICATION_JSON)
                        .accept(MediaType.APPLICATION_JSON)
                )
                .andDo(print())
                .andExpect(status().isCreated());
    }

    @Test
    void shouldGetCustomerSuccessfully() throws Exception {
        when(customerService.getCustomer("REF_001"))
                .thenReturn(CustomerProducer.customerDto(UUID.randomUUID()));
        this.mockMvc.perform(get("/customers/{id}", "REF_001")
                        .accept(MediaType.APPLICATION_JSON))
                .andDo(print())
                .andExpect(status().isOk())
                .andExpect(content().contentType(MediaType.APPLICATION_JSON));
    }

    @Test
    void shouldReturnCustomerNotFound() throws Exception {
        when(customerService.getCustomer("REF_001"))
                .thenThrow(new CustomerNotFoundException("Customer not found of id REF_001"));

        this.mockMvc.perform(get("/customers/{id}", "REF_001")
                .accept(MediaType.APPLICATION_JSON))
                .andDo(print())
                .andExpect(status().isNotFound())
                .andExpect(content().string(Matchers.
                        containsString("Customer not found of id REF_001")))
                .andExpect(content().contentType(
                        MediaType.APPLICATION_JSON));
    }

    public static String asJsonString(final Object obj) {
        try {
            return new ObjectMapper().writeValueAsString(obj);
        } catch (Exception e) {
            throw new RuntimeException(e);
        }
    }

}
